+++ basketball/conf/evolutions/default/1.sql
@@ -151,8 +151,8 @@ create table standing (
   winPercentage             float not null,
   pointDifferential         smallint not null,
   pointDifferentialPerGame  float not null,
-  sumOpptWins               smallint not null,
-  sumOpptGamesPlayed        smallint not null,
+  sumOpptWins               integer not null,
+  sumOpptGamesPlayed        integer not null,
   constraint ck_standing_streakType check (streakType in ('Loss','Win')),
   constraint pk_standing primary key (id))
 ;
